Feat : 기존 테스트 스크립트 보정 (Promehteus 활용하도록
